+++ routes/api.php
@@ -13,37 +13,39 @@ use Illuminate\Http\Request;
 |
 */
 
-// Route::middleware('auth:api')->get('/user', function (Request $request) {
-//     return $request->user();
-// });
-
-Route::prefix('SMH')->group(function () {
-    Route::get('search/{name}', 'Api\SMHPluginController@search');
-    Route::get('book/{bid}', 'Api\SMHPluginController@book');
-    Route::get('book/{bid}/chapter/{cid}', 'Api\SMHPluginController@chapter');
-});
-
 Route::group(['prefix' => 'auth'], function() {
     Route::post('register', 'Api\RegisterController@register');
     Route::post('login', 'Api\Auth\LoginController@login');
     Route::get('logout', 'Api\Auth\LoginController@logout')->middleware('jwt.auth');
 });
 
-Route::get('user', 'Api\UserController@index')->middleware('jwt.auth');
-
-// Content-Type : application/x-www-form-urlencode
-Route::group(['prefix' => 'favorites','middleware', 'jwt.auth'], function() {
-    Route::get('', 'Api\FavoriteController@index');
-    Route::get('/{id}', 'Api\FavoriteController@show');
-    Route::post('', 'Api\FavoriteController@store');
-    Route::put('/{id}', 'Api\FavoriteController@update');
-    Route::delete('/{id}', 'Api\FavoriteController@delete');
+Route::group([], function() {
+    Route::get('users', 'Api\UserController@index')->middleware('jwt.auth');
+    Route::get('users/{id}', 'Api\UserController@show')->middleware('jwt.auth');
+    // Route::post('users', 'Api\RegisterController@register');    // register
+    // Route::put('users/{}', 'Api\Auth\LoginController@login')
+    // Route::delete('users/{}', 'Api\Auth\LoginController@logout')->middleware('jwt.auth');
 });
 
-Route::group(['prefix' => 'resources'], function() {
-    Route::get('', 'Api\ResourceController@index');
-    Route::get('/{id}', 'Api\ResourceController@show');
-    Route::post('', 'Api\ResourceController@store');
-    Route::put('/{id}', 'Api\ResourceController@update');
-    Route::delete('/{id}', 'Api\ResourceController@delete');
+// Content-Type : application/x-www-form-urlencode
+Route::group(['middleware', 'jwt.auth'], function() {
+    Route::get('users/{uid}/favorites', 'Api\FavoriteController@index');
+    Route::get('users/{uid}/favorites/{id}', 'Api\FavoriteController@show');
+    Route::post('users/{uid}/favorites', 'Api\FavoriteController@store');
+    Route::put('users/{uid}/favorites/{id}', 'Api\FavoriteController@update');
+    Route::delete('users/{uid}/favorites/{id}', 'Api\FavoriteController@delete');
+});
+
+Route::group([], function() {
+    Route::get('resources', 'Api\ResourceController@index');
+    Route::get('resources/{id}', 'Api\ResourceController@show');
+    Route::post('resources', 'Api\ResourceController@store');
+    Route::put('resources/{id}', 'Api\ResourceController@update');
+    Route::delete('resources/{id}', 'Api\ResourceController@delete');
+});
+
+Route::prefix('SMH')->group(function () {
+    Route::get('search/{name}', 'Api\SMHPluginController@search');
+    Route::get('book/{bid}', 'Api\SMHPluginController@book');
+    Route::get('book/{bid}/chapter/{cid}', 'Api\SMHPluginController@chapter');
 });
